feat: Add verbose log message when starting file scan
This commit adds a log message to indicate when the script starts scanning
files for markdown documents.

This message is only displayed when verbose mode is enabled using the `-v`
or `--verbose` flags.

This provides users with feedback that the script is actively working,
especially in larger note directories where scanning might take a noticeable
amount of time.